My first attempt to resolve roxygen issue 261
If eval(help_call) cannot find the help for the topic, it raises an
error condition, which eventually causes devtools::document() to fail.
This is my first attempt to use tryCatch, and my first attempt at
handling this error and returning something that will do what I think is
needed to keep devtools::document() running as it should.