add functions to toggle use of doxygen in a given package
add_doxygen_to_roclets and rm_doxygen_from_roclets allow
to switch on and off the use of doxygen as a default
roclet, for example when running document().
It changes the options of Roxygen in the DESCRIPTION file.

In addition the use_doxygen function sets up doxygen
structure and switch on the use of doxygen.