Markdown: don't escape leading ws, \preformatted is protected
Ideally we should keep leading ws in general, because that's what
Roxygen currently does. But it is also tricky, because the
ws might be followed by an ordered or unordered list, or a >
for a block quote, etc.

The parsed XML does not have the ws any more, so we need to
handle it before the markdown run.