Modify ignore_files() for empty .Rbuildignore (#646)
* Modify ignore_files() for empty .Rbuildignore

If the .Rbuildignore file is empty, or more generally if all of the lines in the
file are empty, then the `ignore_files` function in utils.R (line 106) will
throw an error rather than simply returning its first argument to the function,
`rfiles`, which is the desired behavior.  In more detail, if the .Rbuildignore
file is empty, then on line 117

  patterns <- patterns[patterns != ""]

returns a length-0 vector, which in turn causes

  matches <- lapply(patterns, grepl, rfiles_relative, perl = TRUE)

to return a length-0 list, which in turn causes

  matches <- Reduce("|", matches)

to return NULL.  This causes the !matches expression in

  rfiles[!matches]

to throw an error.  The proposed modification simply checks whether
patterns[patterns != ""] is a length-0 vector, and if so returns the first
argument to the function, `rfiles`.

* Remove unneeded branch in ignore_files()

Removed the branch call that came from an else statement that came after an
`if (condition) return(val)` statement.

* Update news to include PR #647

* Modify NEWS.md spaces to prevent merge conflict

There were trailing spaces in the upstream branch in NEWS.md that were nuked in
the previoud commit.  This commit reverts those changes so that the PR doesn't
cause a merge conflict.