Set width argument of wrapString() call to 75 for usage section
At present, it's set to 80. Given that a further 5 space indent is
added to the Usage section in the process of compiling rendered help
files from Rd files, this means that the section can have width up to
85. This causes unattractive/distracting wrap around of some lines in
textual help output (i.e. that got by doing `help(.,
help_type="text")`) in devices whose default viewing window is 80
characters wide.

Not sure that my proposed solution is optimal, but it involves few
enough lines of code that it should be easy to review.