Stop quickly if DESCRIPTION not found
Add an early check to first_time() to see if a DESCRIPTION file exists in the given folder. If not, stop with an informative error message. Previously, a `man` directory was already created in the given directory before an attempt to update the DESCRIPTION file in roxygen_setup() failed with an uninformative "file not found" error.