Refactor writing Win/Unix line endings
We need to move the detection of line endings
upstream a bit, to be able to tell if the
newly generated contents is different from the
current one, in write_if_different().

This means that we need to paste() the contents
together in write_if_different(). Consequently,
we might as well remove the paste()-ing from
same_contents(), as write_if_different() always
calls it with the paste()-d contents.

Also, if we paste() already upstream, then
we might as well use writeBin() to write out
the contents, instead of write_lines(), especially
because write_lines() would add an extra empty line.
write_lines() is still kept, and used in some
code paths, and it can also do EOL-detection,
although this feature is not used currently.

Another benefit of this refactoring is that we
also move enc2utf8() upstream, this should clearly
happen before we check if the new contents is
different.

Added test cases as well.

Fixes #989, supersedes #1003.