Allow restricting image formats
This is currently opt in. If you set `restrict_image_formats`
to `TRUE`, then SVG images are HTML-only and PDF images are
PDF-only. Other file types are not restricted.

We could try to make it opt out, because SVGs do not work in
LaTeX and PDFs most of the time do not work in HTML.

If it is opt in, then we don't necessarily need a NEWS item
to try this out, but if it ends up being opt out, then we
do and we also need to document it,

Closes #1398.