Use cli for hashing
Because it works on Windows network drive paths.
This uses a SHA256 hash, instead of MD5, because MD5 is only
implemented in dev cli for files, but for the file sizes we
use this does not matter. `cli::hash_file_sha256()` is almost
as fast as `digest::digest()` for MD5, actually.

We don't use `tools::md5sum()` because it also has an issue
with Windows paths:
https://bugs.r-project.org/show_bug.cgi?id=17633
Plus we would also need a way to hash strings.

Another alternative would be `rlang::hash_file()`, but rlang
also does not currently has a function to hash strings.

Closes #1443.